+++ Program.cs
@@ -1,6 +1,8 @@
 ﻿using LabManager.Database;
+using LabManager.Models;
 using LabManager.Repositories;
 using Microsoft.Data.Sqlite;
+
 
 var databaseConfig = new DatabaseConfig();
 new DatabaseSetup(databaseConfig);
@@ -26,21 +28,14 @@ if(modelName == "Computer")
 
     if(modelAction == "New")
     {
-         var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
+        Console.WriteLine("Computer New");
+        var id = Convert.ToInt32(args[2]);
+        var ram = args[3];
+        var processor = args[4];
 
-        int id = Convert.ToInt32(args[2]);
-        string ram = args[3];
-        string processor = args[4];
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-        command.ExecuteNonQuery();
-
-        connection.Close();
-    }
+    }   
 }
